+++ src/core/plugin-manager.ts
@@ -4,12 +4,15 @@ import { Client } from './client';
 import { environment } from './../models/environment';
 import fs = require('fs');
 import { Storage } from '../services/storage';
+import { IPluginInfo } from './../models/plugin-info';
 
 const PLUGIN_REGEX = /^.+\.js$/;
 
 export class PluginManager {
 
     static loadPlugins(): void {
+        this.pluginInfo = [];
+        this.pluginInstances = {};
         const folderPath = Storage.makePath('dist', 'plugins');
         let files: string[] = [];
         try {
@@ -27,18 +30,18 @@ export class PluginManager {
                     continue;
                 }
                 const pluginClass = require(relPath).default;
-                const plugin = new pluginClass();
-                const type = (plugin as object).constructor.name;
-                if (!this.pluginInstances) {
-                    this.pluginInstances = {};
-                }
-                this.pluginInstances[type] = plugin;
             } catch (err) {
                 Log('PluginManager', 'Error while loading ' + files[i], LogLevel.Warning);
                 if (environment.debug) {
                     Log('PluginManager', err);
                 }
             }
+        }
+        if (this.afterInitFunctions && this.afterInitFunctions.length > 0) {
+            for (let i = 0; i < this.afterInitFunctions.length; i++) {
+                this.afterInitFunctions[i]();
+            }
+            this.afterInitFunctions = null;
         }
     }
 
@@ -55,12 +58,48 @@ export class PluginManager {
         });
     }
 
-    static addPlugin(info: { name: string, author: string, description?: string }): void {
-        if (!this.pluginInfo) {
-            this.pluginInfo = [];
+    static addPlugin(info: IPluginInfo, target: new () => object): void {
+        // if the plugin is disabled, don't load it.
+        if (info.hasOwnProperty('enabled')) {
+            if (!info.enabled) {
+                Log('PluginManager', 'Skipping disabled plugin ' + info.name, LogLevel.Info);
+                // remove hooks
+                const hKeys = Object.keys(this.hooks);
+                for (let i = 0; i < hKeys.length; i++) {
+                    this.hooks[+hKeys[i]] = this.hooks[+hKeys[i]].filter((hook) => {
+                        return hook.caller !== target.name;
+                    });
+                }
+                return;
+            }
         }
-        Log('PluginManager', 'Loaded ' + info.name + ' by ' + info.author, LogLevel.Info);
+        let plugin: object;
+        try {
+            plugin = new target();
+        } catch (error) {
+            Log('PluginManager', 'Error while instantiating ' + target.name, LogLevel.Warning);
+            if (environment.debug) {
+                Log('PluginManager', error);
+            }
+            return;
+        }
+        this.pluginInstances[target.name] = plugin;
         this.pluginInfo.push(info);
+        Log('PluginManager', 'Loaded ' + info.name + ' by ' + info.author, LogLevel.Info);
+    }
+
+    static getInstanceOf<T extends object>(instance: new () => T): T | null {
+        if (!this.pluginInstances.hasOwnProperty(instance.name)) {
+            return null;
+        }
+        return this.pluginInstances[instance.name] as T;
+    }
+
+    static afterInit(method: () => void): void {
+        if (!this.afterInitFunctions) {
+            this.afterInitFunctions = [];
+        }
+        this.afterInitFunctions.push(method);
     }
 
     static callHooks(packetType: PacketType, packet: Packet, client: object): void {
@@ -87,6 +126,7 @@ export class PluginManager {
     }
 
     private static hooks: { [id: number]: Array<{ action: typeof Function, caller: string }> };
-    private static pluginInfo: Array<{ name: string, author: string, description?: string }>;
-    private static pluginInstances: { [id: string]: object };
+    private static pluginInfo: IPluginInfo[];
+    private static pluginInstances: { [name: string]: object };
+    private static afterInitFunctions: Array<() => void>;
 }