+++ src/core/plugin-module.ts
@@ -4,3 +4,4 @@ export * from './../decorators/plugin';
 export * from './../networking/packets';
 export * from './../networking/packet';
 export * from './../services/logger';
+export * from './plugin-manager';