+++ src/decorators/plugin.ts
@@ -1,10 +1,11 @@
 import { PacketType } from './../networking/packet';
 import { Log, LogLevel } from './../services/logger';
 import { PluginManager } from './../core/plugin-manager';
+import { IPluginInfo } from './../models/plugin-info';
 
-export function NrPlugin(info: { name: string, author: string, description?: string }): (target: object) => any {
-    return (target: object) => {
-        PluginManager.addPlugin(info);
+export function NrPlugin(info: IPluginInfo): (target: new () => object) => any {
+    return (target: new () => object) => {
+        PluginManager.addPlugin(info, target);
         return target;
     };
 }